refactor: improve NetworkConnection service
